Fixing errors when saving single and multiple line items

+++ src/main/java/edu/ucdavis/dss/ipa/repositories/LineItemRepository.java
@@ -22,6 +22,6 @@ public interface LineItemRepository extends CrudRepository<LineItem, Long> {
 
     @Modifying
     @Query("delete from LineItem l where l.id in ?1")
-    void deleteLineItemsWithIds(List<Integer> ids);
+    void deleteLineItemsWithIds(List<Long> ids);
 
 }
