+++ src/main/java/edu/ucdavis/dss/ipa/services/LineItemService.java
@@ -22,7 +22,7 @@ public interface LineItemService {
 
     LineItem createDuplicate(LineItem originalLineItem, BudgetScenario budgetScenario);
 
-    void deleteMany(List<Integer> lineItemIds);
+    void deleteMany(List<Long> lineItemIds);
 
     void createLineItemFromTeachingAssignmentAndBudgetScenario(TeachingAssignment teachingAssignment, BudgetScenario budgetScenario);
 }
